feat: 정기 푸시알림(테스트용: 1분) 전송 로직 추가

+++ src/main/java/com/booksy/domain/notification/repository/DeviceTokenRepository.java
@@ -9,5 +9,5 @@ public interface DeviceTokenRepository extends JpaRepository<DeviceToken, Long> 
 
   Optional<DeviceToken> findByToken(String token);
 
-  List<DeviceToken> findAllByUserId(Long userId);
+  List<DeviceToken> findAllByUserId(Integer userId);
 }
